Message to Professor and TAs:
This is my final version and ready for grading.

As I progressed to Problems 3 and 4 I was able to make improvements to the code that I ran in Problem 2 but did not go back and update the Problem 2 code. You may see some variations in the way I approached the requirements due to this.

Also, I am getting an error that states "Scene is unreachable due to lack of entry points". I have searched to try to figure out why I am getting this and tried a few different things but cannot for the life of me figure out how to fix this. I doesn't seem to affect the operation and flow of the app so I suppose I am risking taking a hit of 3 points due to this, but with an hour before due time, I don't want to potentially break something bigger and not have a larger component fail. Hopefully since the error isn't, at least visibly, disrupting anything, I hope maybe it won't be a full mark down.

I also commented to help walk through my logic and flow, but if it is not helpful I can discontinue commenting in the future, but JP mentioned comments are helpful.