Final Push for Grading  (Van advised in class he would extend the deadline to midnight, so hopefully this is not considered late) - I may submit tomorrow with some enhancements but I will send an email to confirm whether a new one will need to be pulled
For the Main.storyboard, I was editing everything in Any/Regular to optimize for iPhone 6 Portrait. I did not have any warnings with this, however it seemed to pull some warning when not viewing in this mode. Hopefully that will not warrant markdowns since the expectation was to have it look nice in iPhone 6 Portrait.